move routes.cs file to Sample.Tests - minor code refactor
